Saving: Update saving strategy
I've updated the way the plugin handles saving annotations. Most
notably, instead of saving *all* the annotations in memory upon quitting
vim, we save a single annotation once the user quits the buffer that
contains the annotation. Eventually we'll want to save the annotations
on successful writes only.

Additionally I've added additional structure to the representation of
signs. This makes them much easier to work with. Sign entries also
contain a reference to the source file they annotate.

Hopefully this change can help push us toward not maintaining *all* the
annotations in memory at once. Since we now append annotations to the
config file we'll also eventually have to check for duplicates,
presumably using the annotation ids. If an annotation with a given id
already exists as an entry in the config file, we should overwrite the
line in the config rather than appending a new one.