List: Add support for listing annotations
Added a window for listing available annotations for the current file.
The implementation is not as clean as it could be, but it presently
works.

* Added a window options function to set the appropriate options for the
list window.
* Added configuration options for the list window.
* Added a single key mapping for the list window (can use the dict and
the same scheme to support others)

This change also necessitated a few minor modifications to the
gadamer#Read function--we need to pass along the line number information
from an annotation listing in the list window, so that's supported.

Also fixed a couple of tiny typos/bugs.