Core: Use window mode functions.
Where appropriate, I've changed calls to splits/windowing to using the
window utility functions in gadamer/window.vim. This makes the overall
mechanics of the plugin clearer and easier to comprehend.

Later on, we'll ensure there's only a one way link between windowing
fucntions and the rest of the plugin. To do so, we'll probably have to
rethink the way we handle window key mappings--often we want to map to
some functionality in another part of the plugin, but the window should
have no knowledge of the other component.