Annotations: Update annotation representations
I've refined the approach to representing and working with annotations.
Plenty of aspects of the implementation are still less than ideal--for
instance, there is a question as to whether the current approach is very
efficient, and whether or not better boundaries might be established in
the annotations dict--e.g. what ought to be the responsibility of the
underlying set and what ought to be exposed by the dict itself.