Window: Define mode object, add view mapping.
I've defined a "class" for window modes to instantiate, which should
help standardize their functionality. This also enabled me to refactor
the `setWindowMappings` function so that the same pattern may be
followed for all calls in the `open` function. Additionally, I should
add the invoke and 'close' functions to the prototype object.

Windows now keep information on the buffer that was active prior to
their being opened--this is necessary in some circumstances.

Added a mapping to the view mode to reopen the previous buffer on close.