Window/Mode: Refactor approach to modes
This change contains a significant refactor of the approach to "modes"
or different windows/user experiences for working with annotations. I've
broken up the modes into separate files. Each mode is now an
instantiation of an abstract definition of a mode, which ensures they
have some set of consistent functionality. In general, this should make
them easier to work with. This should also make spinning up new modes
simpler too. I've also updated the core script to use the new approach.