Handle index bounds errors in list mode
In list mode, we need to manually update the cursor position in order to
overload vim's normal functionality with gadamer specific functions.
There was an unhandled case in which we attempted to gather an
annotation from the first line in the list window, which is blank. This
change handles that error.