Update gadamer.vim according to sign api changes
This change updates calls to the sign mgmt api to reflect recent
changes. Most significantly, we now open an annotation before placing a
sign, and use the fromAnnotation function where appropriate. We also
store a reference to the file buffer before placing a sign, since the
default behavior of sign placement is to place the sign in the current
buffer (which, after opening an annotation window, would be the
annotation itself, not the associated file).

Additionally, I've also updated the signatures of the main API to set
the current line as the default argument (applies to Read and Annotate).

I've also removed trailing whitespace.