Add support for range based annotations
This commit enhances annotations so that we may define them across
ranges of lines in a file.

I've had to update several call sites to support the ranged variant of
an annotation. If an annotation falls on a single line, it's range is
the same line (e.g. 1,1).

The annotations set is now keyed by *starting lines* which each are sets
in themselves, keyed by *ending lines*. Thus, a specification like `1,1
3,3 3,5` yields the following data structure:

```vimscript
{ '1': { '1': {...annotation content}},
  '3': { '3': {...annotation}, '5': {...annotation}}}
```

Thus, annotations, which were previously unique across lines, are now
unique across ranges.