Fix annotations set membership
The second clause of the logical and returned by the membership function
could fail when the starting line isn't a member of the set.

We now perform each check individually to rectify the situation--we were
also checking for the wrong property in the second clause, originally
(end instead of start).