Add visualizer mode
This change adds a new mode, "visualizer" which opens a visual map of
annotation ranges in a window adjacent to the main file.

I had to make several changes to the mode object and methods to make
this work, primarily refactoring how we handle help displays.

I've also updated tools and the main autoload file accordingly.